requirements.txt
google-generativeai==0.8.3
gTTS==2.5.4
SpeechRecognition==3.11.0